# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FPLense API",
    "summary": "FPL points forecasts, squad optimizer, season replay and team tools.",
    "description": "Prices are integer tenths of £1m (55 = £5.5m). Expected points come from a LightGBM model validated walk-forward; hindsight-best squads are the best legal £100m squads on actual points (not FPL's Dream Team). Not affiliated with the Premier League.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/players": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Players",
        "description": "Every player for a GW: price, availability, rating, photo URL, expected points per GW,\n``P_h``, actual points once the GW is finished, and the next 5 GWs' fixtures with FDR.",
        "operationId": "players_api_players_get",
        "parameters": [
          {
            "name": "gw",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 38,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "default: the next gameweek",
              "title": "Gw"
            },
            "description": "default: the next gameweek"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayersOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/players/{player_id}": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Player",
        "description": "One player: expected vs actual points each GW this season, the next-GW SHAP waterfall\n(\"why the model thinks this\"), fixtures and ownership.",
        "operationId": "player_api_players__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/entry/{team_id}": {
      "get": {
        "tags": [
          "entry"
        ],
        "summary": "Entry",
        "description": "Team and manager name, rank, value and bank, and every GW's points, rank, transfers,\nhits and chip, next to the model pick, the hindsight-best and the average score.",
        "operationId": "entry_api_entry__team_id__get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 100000000,
              "exclusiveMinimum": 0,
              "description": "the number in your FPL team URL",
              "title": "Team Id"
            },
            "description": "the number in your FPL team URL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/entry/{team_id}/gw/{gw}": {
      "get": {
        "tags": [
          "entry"
        ],
        "summary": "Entry Gw",
        "description": "That GW's picks resolved to players, with expected points (from the frozen archive) and\nactual points, plus the lineup helper's best XI and captain from the same 15.",
        "operationId": "entry_gw_api_entry__team_id__gw__gw__get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 100000000,
              "exclusiveMinimum": 0,
              "description": "the number in your FPL team URL",
              "title": "Team Id"
            },
            "description": "the number in your FPL team URL"
          },
          {
            "name": "gw",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 38,
              "minimum": 1,
              "title": "Gw"
            }
          },
          {
            "name": "predict_gw",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 38,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "GW the lineup helper optimises (default: this GW if predictions exist for it, else the next GW)",
              "title": "Predict Gw"
            },
            "description": "GW the lineup helper optimises (default: this GW if predictions exist for it, else the next GW)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryGwOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/codes/decode": {
      "post": {
        "tags": [
          "codes"
        ],
        "summary": "Decode",
        "description": "A team code (or a share URL containing one) -> the squad with expected points for the\nnext GW. Bad codes answer 422 with a specific message (typo, other season, illegal team).",
        "operationId": "decode_api_codes_decode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecodeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/codes/encode": {
      "post": {
        "tags": [
          "codes"
        ],
        "summary": "Encode",
        "description": "11 starters + 4 bench (in order) + captain and vice -> a shareable code.",
        "operationId": "encode_api_codes_encode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EncodeIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncodeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DecodeIn": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 500,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "DecodeIn"
      },
      "DecodeOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "season": {
            "type": "string",
            "title": "Season"
          },
          "squad": {
            "$ref": "#/components/schemas/Squad"
          },
          "over_budget": {
            "type": "boolean",
            "title": "Over Budget"
          }
        },
        "type": "object",
        "required": [
          "code",
          "season",
          "squad",
          "over_budget"
        ],
        "title": "DecodeOut"
      },
      "EncodeIn": {
        "properties": {
          "starters": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 11,
            "minItems": 11,
            "title": "Starters"
          },
          "bench": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Bench"
          },
          "captain": {
            "type": "integer",
            "title": "Captain"
          },
          "vice": {
            "type": "integer",
            "title": "Vice"
          }
        },
        "type": "object",
        "required": [
          "starters",
          "bench",
          "captain",
          "vice"
        ],
        "title": "EncodeIn"
      },
      "EncodeOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "share_param": {
            "type": "string",
            "title": "Share Param"
          }
        },
        "type": "object",
        "required": [
          "code",
          "share_param"
        ],
        "title": "EncodeOut"
      },
      "Entry": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "team_name": {
            "type": "string",
            "title": "Team Name"
          },
          "manager_name": {
            "type": "string",
            "title": "Manager Name"
          },
          "overall_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Points"
          },
          "overall_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Rank"
          },
          "current_event": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Event"
          },
          "value": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "bank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bank"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/EntryGw"
            },
            "type": "array",
            "title": "History"
          },
          "chips": {
            "items": {
              "$ref": "#/components/schemas/Chip"
            },
            "type": "array",
            "title": "Chips"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_name",
          "manager_name",
          "overall_points",
          "overall_rank",
          "current_event",
          "value",
          "bank",
          "history",
          "chips"
        ],
        "title": "Entry"
      },
...
      "EntryGwOut": {
        "properties": {
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "gw": {
            "type": "integer",
            "title": "Gw"
          },
          "predict_gw": {
            "type": "integer",
            "title": "Predict Gw"
          },
          "active_chip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Chip"
          },
          "points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "bank": {
            "type": "integer",
            "title": "Bank"
          },
          "value": {
            "type": "integer",
            "title": "Value"
          },
          "squad": {
            "$ref": "#/components/schemas/Squad"
          },
          "automatic_subs": {
            "items": {
              "$ref": "#/components/schemas/AutoSub"
            },
            "type": "array",
            "title": "Automatic Subs"
          },
          "lineup": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Squad"
              },
              {
                "type": "null"
              }
            ],
            "description": "lineup helper: best XI + captain from these 15"
          },
          "lineup_changes": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Lineup Changes",
            "description": "players the helper moves into the XI"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "gw",
          "predict_gw",
          "active_chip",
          "points",
          "bank",
          "value",
          "squad",
          "automatic_subs",
          "lineup",
          "lineup_changes"
        ],
        "title": "EntryGwOut"
      },
...
      "GwPoints": {
        "properties": {
          "gw": {
            "type": "integer",
            "title": "Gw"
          },
          "points": {
            "type": "number",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "gw",
          "points"
        ],
        "title": "GwPoints"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlayerDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "code": {
            "type": "integer",
            "title": "Code"
          },
          "web_name": {
            "type": "string",
            "title": "Web Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "pos": {
            "type": "string",
            "enum": [
              "GK",
              "DEF",
              "MID",
              "FWD"
            ],
            "title": "Pos"
          },
          "club": {
            "type": "string",
            "title": "Club"
          },
          "club_short": {
            "type": "string",
            "title": "Club Short"
          },
          "price": {
            "type": "integer",
            "title": "Price"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "chance_of_playing_next_round": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chance Of Playing Next Round"
          },
          "news": {
            "type": "string",
            "title": "News"
          },
          "availability": {
            "type": "number",
            "title": "Availability"
          },
          "selected_by_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected By Percent"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "rating_source": {
            "type": "string",
            "title": "Rating Source"
          },
          "expected": {
            "items": {
              "$ref": "#/components/schemas/GwPoints"
            },
            "type": "array",
            "title": "Expected",
            "description": "expected points per GW (live horizon)"
          },
          "p": {
            "type": "number",
            "title": "P",
            "description": "expected points in the requested GW"
          },
          "P_h": {
            "type": "number",
            "title": "P H",
            "description": "discounted expected points over the horizon"
          },
          "actual": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual",
            "description": "actual points if the GW is finished"
          },
          "minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Minutes"
          },
          "fixtures": {
            "items": {
              "$ref": "#/components/schemas/FixtureTick"
            },
            "type": "array",
            "title": "Fixtures"
          },
          "season_points": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Points"
          },
          "season_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season Minutes"
          },
          "prev_season_pts_per90": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Season Pts Per90"
          },
          "top_shap": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Shap"
          },
          "shap_base_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shap Base Value"
          },
          "shap_prediction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shap Prediction",
            "description": "raw next-GW prediction before availability"
          },
          "shap": {
            "items": {
              "$ref": "#/components/schemas/ShapItem"
            },
            "type": "array",
            "title": "Shap"
          },
          "season": {
            "items": {
              "$ref": "#/components/schemas/PlayerGw"
            },
            "type": "array",
            "title": "Season"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "web_name",
          "name",
          "pos",
          "club",
          "club_short",
          "price",
          "status",
          "chance_of_playing_next_round",
          "news",
          "availability",
          "selected_by_percent",
          "photo_url",
          "rating",
          "rating_source",
          "expected",
          "p",
          "P_h",
          "fixtures",
          "season_points",
          "season_minutes",
          "prev_season_pts_per90",
          "top_shap",
          "shap_base_value",
          "shap_prediction",
          "shap",
          "season"
        ],
        "title": "PlayerDetail"
      },
...
      "PlayersOut": {
        "properties": {
          "gw": {
            "type": "integer",
            "title": "Gw"
          },
          "gws": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Gws"
          },
          "source": {
            "type": "string",
            "enum": [
              "live",
              "backfill",
              "latest"
            ],
            "title": "Source"
          },
          "finished": {
            "type": "boolean",
            "title": "Finished"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/Player"
            },
            "type": "array",
            "title": "Players"
          }
        },
        "type": "object",
        "required": [
          "gw",
          "gws",
          "source",
          "finished",
          "players"
        ],
        "title": "PlayersOut"
      },
...
      "ShapItem": {
        "properties": {
          "feature": {
            "type": "string",
            "title": "Feature"
          },
          "shap": {
            "type": "number",
            "title": "Shap"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "feature",
          "shap",
          "value"
        ],
        "title": "ShapItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}